feat: component, context, type 분리
